chore(#1472): Split constraint reader atomic constraints
Atomic constraint reading bloats the constraint reader, so I've
separated it out into another file.
