+++ profile/src/test/java/com/scottlogic/deg/profile/reader/JsonProfileReaderTests.java
@@ -77,7 +77,7 @@ public class JsonProfileReaderTests {
     private final String schemaVersion = "\"0.7\"";
     private String json;
 
-    private JsonProfileReader jsonProfileReader = new JsonProfileReader(null, new ConstraintReader(new MockFromFileReader()));
+    private JsonProfileReader jsonProfileReader = new JsonProfileReader(null, new ConstraintReader(new AtomicConstraintReader(new MockFromFileReader())));
